修改 tcp例程

+++ network/tcp.h
@@ -5,8 +5,8 @@
 class TCPCLIENT
 {
 	public:
-		u8 begin(SOCKET ps,uint16 port);
-		u8 connect(u8* IP,uint16 Port);
+		int begin(SOCKET ps,uint16 port);
+		int connect(u8* IP,uint16 Port);
 		u16 recv(u8* buf);
 //		u16 interruptRecv(u8* buf);
 	  u16 send(u8* buf,u16 len);
